mcp-server/requirements.txt
fastmcp>=2.0.0,<3.0.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
PyYAML>=6.0.0
